backend/requirements.txt
aiohttp
python-dotenv
fastapi
uvicorn[standard]
aiofiles
apscheduler
pydantic[email]
python-multipart
pyjwt
orjson
bs4
nltk
pandas
pyarrow
redis
aioredis==1.3.1
google-cloud-secret-manager
torch
transformers
tqdm
symspellpy
chardet
python-Levenshtein
pydantic
gensim